In this commit
1. We allow the users to enter Servings quantities with two new units: litres and kilograms

2. We update the `toAtom(r: Recipe, cr: CuratedRecipe): Atom` function with an automatic conversion of
   litres and kilograms ( quantities and units ) to millilitres and grams

3. Note that the return value of atomServeQuantity is always Short
   therefore, in the case of millilitres and grams, we convert the Double to Short

The strings "litres", "kilograms", "millilitres" and "grams" will be made into a type in the next PR :)

+++ common/src/main/scala/com/gu/recipeasy/models/Recipe.scala
@@ -53,7 +53,7 @@ case class IngredientsList(
 )
 
 case class Serves(
-  from: Int,
-  to: Int
+  from: Double,
+  to: Double
 )
